Add code_extractor_reinsert_test.rb
This extracts the "unextract" (aka: "reinsert") test into it's own test,
with the intent to later fill out other test cases in future commits.

Note:  The `create_base_repo` needed to be added to the `test_helper.rb`
since it was needed by both test files.  Not a huge change, but worth
nothing.
